temp_scripts/requirements_perspective.txt
pandas>=1.5.0
numpy>=1.21.0
google-api-python-client>=2.0.0
tqdm>=4.65.0 